+++ cvtModel/src/models/ramps/linear_segment.py
@@ -1,4 +1,4 @@
-from utils.ramps.ramp_segment import RampSegment
+from models.ramps.ramp_segment import RampSegment
 
 
 class LinearSegment(RampSegment):
