+++ cvtModel/src/models/ramps/pro_defined_segment.py
@@ -1,6 +1,6 @@
 import math
 from constants.car_specs import INITIAL_FLYWEIGHT_RADIUS
-from utils.ramps.ramp_segment import RampSegment
+from models.ramps.ramp_segment import RampSegment
 
 
 class ProDefinedSegment(RampSegment):
