+++ cvtModel/src/models/ramps/ramp_segment.py
