+++ cvtModel/src/utils/ramp_utils.py
